Resolved warning

+++ src/graph.cpp
@@ -326,7 +326,7 @@ void print_circuit(NODE *graph, int num_of_nodes)
     char title[] = "| ADDRESS\t | NAME\t | TYPE\t | PRIMARY OUTPUT #\t | INPUT #\t | OUTPUT #\t | CORRECT VALUE\t | FAULT VALUE\t | MARKER\t | FANIN\t | FANOUT |";
     char msg[512];
 
-    sprintf(msg, "%0*d", strlen(title - 2), 0);
+    sprintf(msg, "%0*d", (int)strlen(title - 2), 0);
     subst(msg, '0', '-');
     prepend(msg, (char *)'+');
     msg[strlen(title - 1)] = '+';
